requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.6
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
nextcord.py==2.5.2
frozenlist==1.6.0
idna==3.10
multidict==6.4.4
propcache==0.3.1
python-dotenv==1.1.0
typing_extensions==4.13.2
yarl==1.20.0
